+++ CHANGELOG.md
@@ -19,6 +19,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fixed an error in `src/Shared/GeosUtil/hco_regrid_a2a_mod.F90` where an accumulator was uninitialized before reuse, which may inherit junk data from the previous iteration if the southmost source cell is not found
 - Fixed IF-block logic errors in `SrcFile_Parse` that led to incorrect time-cycling behavior
+- Fixed an error in `src/Shared/GeosUtil/hco_regrid_a2a_mod.F90` where `ymap` southern edge straddling the source cell would be `missval` incorrectly.
+- Fixed an error in `src/Shared/GeosUtil/hco_regrid_a2a_mod.F90` where `xmap` ghost regions were initialized as `0` and not `missval`.
+- Fixed an error in `src/Shared/GeosUtil/hco_regrid_a2a_mod.F90` where `ymap_r8r8` incorrectly uses `im` as the divisor instead of `nlon`; and `sum` is calculated based on `nlon > 0.0` but the assignment of `q2 = sum` is not in that `if` statement leading to clobbered data.
 
 #### Removed
 - Removed C-preprocessor switch `ESMF_`